Add Translate (#4)
